# openAPI.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Dataloaders API",
        "description": "API documentation for Dataloaders application",
        "contact": {
            "name": "Dataloaders Team",
            "email": "support@dataloaders.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8083",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/v1/dataloader/deploy/{deployId}": {
            "get": {
                "tags": [
                    "Deploy"
                ],
                "summary": "Get deployment",
                "description": "Get deployment by ID",
                "operationId": "getDeploy",
                "parameters": [
                    {
                        "name": "deployId",
                        "in": "path",
                        "description": "Deploy ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deployment found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Deploy"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Deployment not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Deploy"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Deploy"
                ],
                "summary": "Update deployment",
                "description": "Update an existing deployment",
                "operationId": "updateDeploy",
                "parameters": [
                    {
                        "name": "deployId",
                        "in": "path",
                        "description": "Deploy ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Deploy"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Deployment updated successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Deploy"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Deployment not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Deploy"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Deploy"
                ],
                "summary": "Delete deployment",
                "description": "Delete a deployment",
                "operationId": "deleteDeploy",
                "parameters": [
                    {
                        "name": "deployId",
                        "in": "path",
                        "description": "Deploy ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deployment deleted successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Deployment not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/dataloader/job/deploy/{jobId}": {
            "post": {
                "tags": [
                    "Job Config"
                ],
                "summary": "Deploy job",
                "description": "Deploy a job configuration",
                "operationId": "deploy",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job deployed successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/dataloader/deploy": {
            "post": {
                "tags": [
                    "Deploy"
                ],
                "summary": "Create deployment",
                "description": "Create a new deployment",
                "operationId": "createDeploy",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Deploy"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Deployment created successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Deploy"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Deploy"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/dataloader/deploy/runningConfig/{jobId}": {
            "get": {
                "tags": [
                    "Deploy"
                ],
                "summary": "Get job config by job ID",
                "description": "Get running job configuration by job ID",
                "operationId": "getJobConfigById",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job config retrieved",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dataloader/deploy/runningConfig/scheduleId/{schedulerId}": {
            "get": {
                "tags": [
                    "Deploy"
                ],
                "summary": "Get job config by scheduler ID",
                "description": "Get running job configuration by scheduler ID",
                "operationId": "getJobConfigBySchedulerId",
                "parameters": [
                    {
                        "name": "schedulerId",
                        "in": "path",
                        "description": "Scheduler ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job config retrieved",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dataloader/deploy/job/{jobId}": {
            "get": {
                "tags": [
                    "Deploy"
                ],
                "summary": "Get deployments by job ID",
                "description": "Get all deployments for a job",
                "operationId": "getDeploysByJobId",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deployments retrieved",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Deploy"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dataloader/deploy/app/{applicationId}": {
            "get": {
                "tags": [
                    "Deploy"
                ],
                "summary": "Get deployments by application",
                "description": "Get all deployments for an application",
                "operationId": "getAllDeploys",
                "parameters": [
                    {
                        "name": "applicationId",
                        "in": "path",
                        "description": "Application ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deployments retrieved",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Deploy"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorDefinition": {
                "type": "object",
                "properties": {
                    "label": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "level": {
                        "type": "string",
                        "enum": [
                            "REQUEST",
                            "BUSINESS",
                            "SECURITY",
                            "CONFLICT",
                            "RESOURCE"
                        ]
                    },
                    "severity": {
                        "type": "string",
                        "enum": [
                            "FATAL",
                            "NONFATAL"
                        ]
                    },
                    "message": {
                        "type": "string"
                    },
                    "httpStatus": {
                        "type": "string",
                        "enum": [
                            "100 CONTINUE",
                            "101 SWITCHING_PROTOCOLS",
                            "102 PROCESSING",
                            "103 EARLY_HINTS",
                            "103 CHECKPOINT",
                            "200 OK",
                            "201 CREATED",
                            "202 ACCEPTED",
                            "203 NON_AUTHORITATIVE_INFORMATION",
                            "204 NO_CONTENT",
                            "205 RESET_CONTENT",
                            "206 PARTIAL_CONTENT",
                            "207 MULTI_STATUS",
                            "208 ALREADY_REPORTED",
                            "226 IM_USED",
                            "300 MULTIPLE_CHOICES",
                            "301 MOVED_PERMANENTLY",
                            "302 FOUND",
                            "302 MOVED_TEMPORARILY",
                            "303 SEE_OTHER",
                            "304 NOT_MODIFIED",
                            "305 USE_PROXY",
                            "307 TEMPORARY_REDIRECT",
                            "308 PERMANENT_REDIRECT",
                            "400 BAD_REQUEST",
                            "401 UNAUTHORIZED",
                            "402 PAYMENT_REQUIRED",
                            "403 FORBIDDEN",
                            "404 NOT_FOUND",
                            "405 METHOD_NOT_ALLOWED",
                            "406 NOT_ACCEPTABLE",
                            "407 PROXY_AUTHENTICATION_REQUIRED",
                            "408 REQUEST_TIMEOUT",
                            "409 CONFLICT",
                            "410 GONE",
                            "411 LENGTH_REQUIRED",
                            "412 PRECONDITION_FAILED",
                            "413 PAYLOAD_TOO_LARGE",
                            "413 REQUEST_ENTITY_TOO_LARGE",
                            "414 URI_TOO_LONG",
                            "414 REQUEST_URI_TOO_LONG",
                            "415 UNSUPPORTED_MEDIA_TYPE",
                            "416 REQUESTED_RANGE_NOT_SATISFIABLE",
                            "417 EXPECTATION_FAILED",
                            "418 I_AM_A_TEAPOT",
                            "419 INSUFFICIENT_SPACE_ON_RESOURCE",
                            "420 METHOD_FAILURE",
                            "421 DESTINATION_LOCKED",
                            "422 UNPROCESSABLE_ENTITY",
                            "423 LOCKED",
                            "424 FAILED_DEPENDENCY",
                            "425 TOO_EARLY",
                            "426 UPGRADE_REQUIRED",
                            "428 PRECONDITION_REQUIRED",
                            "429 TOO_MANY_REQUESTS",
                            "431 REQUEST_HEADER_FIELDS_TOO_LARGE",
                            "451 UNAVAILABLE_FOR_LEGAL_REASONS",
                            "500 INTERNAL_SERVER_ERROR",
                            "501 NOT_IMPLEMENTED",
                            "502 BAD_GATEWAY",
                            "503 SERVICE_UNAVAILABLE",
                            "504 GATEWAY_TIMEOUT",
                            "505 HTTP_VERSION_NOT_SUPPORTED",
                            "506 VARIANT_ALSO_NEGOTIATES",
                            "507 INSUFFICIENT_STORAGE",
                            "508 LOOP_DETECTED",
                            "509 BANDWIDTH_LIMIT_EXCEEDED",
                            "510 NOT_EXTENDED",
                            "511 NETWORK_AUTHENTICATION_REQUIRED"
                        ]
                    },
                    "sourceApplication": {
                        "type": "string"
                    }
                }
            },
            "Response": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    },
                    "causedBy": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ErrorDefinition"
                        }
                    }
                }
            },
...
            "Deploy": {
                "type": "object",
                "properties": {
                    "createdAt": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "createdBy": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "updatedBy": {
                        "type": "string"
                    },
                    "deployId": {
                        "type": "string"
                    },
                    "parentJobId": {
                        "type": "string"
                    },
                    "jobId": {
                        "type": "string"
                    },
                    "deployName": {
                        "type": "string"
                    },
                    "manualRun": {
                        "type": "boolean"
                    },
                    "schedule": {
                        "type": "boolean"
                    },
                    "scheduleExpression": {
                        "type": "string"
                    },
                    "schedulerId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "schedulerName": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "deletedAt": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
        }
    }
}